use binary strings instead of encode/decode
The encode()/decode() dance is a little noisy. I think it would be better to use binary file opens and strings instead. It would still make it explicit that arguments are binary while at the same time having minimal noise.

I will note that ONE encode does remain in the README, and it might be a sign of a discrepancy in the API: if the READMe is correct, encrypt_bytes takes bytes but returns a string, and decrypt_bytes takes bytes and returns bytes. Should this not be orthogonal?

In general, I would encourage avoiding ASCII-armored text in the API unless explicitely requested. In this case, I would cetainly not expect encypt_bytes to return a string!