Adds encrypt_bytes_to_file function
encrypt_bytes_to_file function takes a list of strings as cert paths,
and data as bytes, and one output file path as bytes to encrypt.

We also fixed one bug in the decryption code, where if the pesk for
the encrypted data is the not the first key who signed it. Say, the
data was encrypted by 10 different keys, and our private key was not
the first key in that list.

Also updated the hello*.asc files for test case.